vendor protobufs, generate Go code (#3)
This repository relied on the generated Go files for the remote execution protos that were in [remote-apis repository](https://github.com/bazelbuild/remote-apis). As per https://github.com/bazelbuild/remote-apis/issues/177, that repo is dropping support for the generated Go files, so vendor the protos and generate the code in the repo.

The Google "well-known protos" are from the [googleapis repository](https://github.com/googleapis/googleapis).