casload: introduce load actions including a read-back action (#7)
This PR introduces the notion of a "load action" to `casload`. Each load action is a type of load test (or an auxillary support action, e.g. load and save the set of "known digests"). The existing write load code becomes a "generate" action. The PR also adds a "read" action to generate read requests from the set of known digests (which can come from the generate action or from the load-digests action).